+++ tests/App/Factory/ContactFactory.php
@@ -5,39 +5,23 @@ declare(strict_types=1);
 namespace araise\TableBundle\Tests\App\Factory;
 
 use araise\TableBundle\Tests\App\Entity\Contact;
-use araise\TableBundle\Tests\App\Repository\ContactRepository;
-use Zenstruck\Foundry\ModelFactory;
-use Zenstruck\Foundry\Proxy;
-use Zenstruck\Foundry\RepositoryProxy;
+use Zenstruck\Foundry\Persistence\PersistentObjectFactory;
 
 /**
- * @method static        Contact|Proxy createOne(array $attributes = [])
- * @method static        Contact[]|Proxy[] createMany(int $number, $attributes = [])
- * @method static        Contact|Proxy find($criteria)
- * @method static        Contact|Proxy findOrCreate(array $attributes)
- * @method static        Contact|Proxy first(string $sortedField = 'id')
- * @method static        Contact|Proxy last(string $sortedField = 'id')
- * @method static        Contact|Proxy random(array $attributes = [])
- * @method static        Contact|Proxy randomOrCreate(array $attributes = [])
- * @method static        Contact[]|Proxy[] all()
- * @method static        Contact[]|Proxy[] findBy(array $attributes)
- * @method static        Contact[]|Proxy[] randomSet(int $number, array $attributes = [])
- * @method static        Contact[]|Proxy[] randomRange(int $min, int $max, array $attributes = [])
- * @method static        ContactRepository|RepositoryProxy repository()
- * @method Contact|Proxy create($attributes = [])
+ * @extends PersistentObjectFactory<Contact>
  */
-final class ContactFactory extends ModelFactory
+final class ContactFactory extends PersistentObjectFactory
 {
-    protected function getDefaults(): array
+    public static function class(): string
+    {
+        return Contact::class;
+    }
+
+    protected function defaults(): array
     {
         return [
             'name' => self::faker()->name(),
             'company' => CompanyFactory::randomOrCreate(),
         ];
-    }
-
-    protected static function getClass(): string
-    {
-        return Contact::class;
     }
 }